requirements.txt
awsebcli==3.6.2
blessed==1.9.5
botocore==1.3.13
cement==2.4.0
cffi==1.3.1
characteristic==14.3.0
colorama==0.3.3
cryptography==1.1.2
cssselect==0.9.1
Django==1.7.11
django-filter==0.11.0
djangorestframework==3.3.2
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
enum34==1.1.1
idna==2.0
ipaddress==1.0.15
jmespath==0.9.0
lxml==3.5.0
Markdown==2.6.5
MySQL-python==1.2.5
pathspec==0.3.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
PyYAML==3.11
queuelib==1.4.2
requests==2.6.2
Scrapy==1.0.3
service-identity==14.0.0
six==1.10.0
texttable==0.8.4
Twisted==15.5.0
w3lib==1.13.0
wcwidth==0.1.5
websocket-client==0.34.0
wheel==0.24.0
zope.interface==4.1.3
